requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
pandas>=2.2.0
rapidfuzz==3.5.2
python-multipart==0.0.6
